REFACTOR: AddSkill, AddMonster로 함수화

+++ ConSoleStudy2023/Skill.h
@@ -4,8 +4,8 @@
 class Skill :
     public Object
 {
-    float durationTime = 0.f;
-    float durationTimeMAX = 10.f;
+    double durationTime = 0.0;
+    double durationTimeMAX = 10.0;
 public:
     Skill();
     Skill(Vec2 Pos, int type, unsigned long long id, int Time);
